# PctBackend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Heures UVCI API",
        "description": "Documentation de l'API de gestion des heures",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Liste de tous les comptes",
                "operationId": "12e989554c281d650dff75443acabf54",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Créer un compte utilisateur (admin ou secrétaire)",
                "operationId": "64da9b5a20fa4f20e947ed61b2bbb6da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password",
                                    "id_profil"
                                ],
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "secretaire2@uvci.edu.ci"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "MonMotDePasse@2024"
                                    },
                                    "id_profil": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "id_enseignant": {
                                        "description": "Obligatoire si le profil est 'enseignant', null sinon",
                                        "type": "integer",
                                        "example": null
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Détail d'un compte",
                "operationId": "a9d321d5dc6bc2613c6cf5ba92b76c89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail utilisateur"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Modifier un compte (login, mot de passe, profil)",
                "operationId": "5bcda8371b3cb9f863a47620e636be9b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "secretaire2@uvci.edu.ci"
                                    },
                                    "password": {
                                        "description": "Laisser vide pour ne pas changer",
                                        "type": "string"
                                    },
                                    "id_profil": {
                                        "type": "integer"
                                    },
                                    "actif": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte modifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Désactiver un compte",
                "operationId": "0446b3779430bbd9ca0f106d9fd0a2ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte désactivé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{id}/reinitialiser-mdp": {
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Réinitialiser le mot de passe d'un utilisateur",
                "operationId": "18dc54a2abb138d69a300830fc860a49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mot de passe réinitialisé, nouveau mdp retourné"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}